+++ components/groupcard.tsx
@@ -1,21 +1,25 @@
 
 import { Pressable, TouchableOpacity, View } from "react-native";
 import { Card, Heading, Image, Text, Button } from "@gluestack-ui/themed";
-import { Link } from "expo-router";
+import { Link, router } from "expo-router";
 
 
-export default function GroupCard() {
-
+export default function GroupCard(props: {name: string, bio: string}) {
         return (
+            <Pressable onPress={() => {
+                router.push({
+                    pathname: "/group/[id]",
+                    params: {id: props.name, bio: props.bio}
+                })
+            }}>
             <Card size="md" variant="elevated" w="$full" >
-
                 <Heading mb="$1" size="md">
-                    Group Name
+                    {props.name}
                 </Heading>
-                <Text size="sm">Group Bio</Text>
-
-            </Card>
+                <Text size="sm">{props.bio}</Text>
+            </Card>         
+            </Pressable>
         );
     
 }
-
+
