+++ package.json
@@ -31,7 +31,7 @@
     "react-dom": "18.2.0",
     "react-native": "0.73.2",
     "react-native-gesture-handler": "~2.14.0",
-    "react-native-reanimated": "~3.6.2",
+    "react-native-reanited": "~3.6.2",
     "react-native-safe-area-context": "^4.9.0",
     "react-native-screens": "~3.29.0",
     "react-native-smooth-pincode-input": "^1.0.9",
